Extract useStatMods hook for shared stat modifier logic
Introduces a useStatMods hook that provides mods, updateMod, addMod,
and removeMod for StatMap editing. Replaces duplicated state management
in RaceStatMods and StatModsEditor components, which both had identical
add/remove/update/cleanup-to-undefined logic.

Closes #16
